Creation of abstract unit test mock variables subpackage
